pyproject.toml
[tool.poetry]
name = "glasses"
version = "0.34.0"
description = ""
authors = ["Sander Teunissen"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
textual = "^0.41.0"
pydantic = "^1.10.2"
kubernetes = "^25.3.0"
kubernetes-asyncio = "^28.2.0"
lark = "^1.1.5"
rich = "^13.3.2"
structlog = "^23.1.0"
textual-dev = "^1.2.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^1.7.0"
ruff = "^0.1.6"
black = "^23.11.0"
pytest-asyncio = "^0.20.3"
pytest-icdiff = "^0.6"
nox = "^2022.11.21"
pyproject-fmt = "^0.4.1"
isort = "^5.12.0"
icdiff = "^2.0.6"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core",
]

[tool.poetry.scripts]
glasses = "glasses.app:run"

[tool.ruff]
src = [
  'src',  # if omitted, Isort messes up sorting of primary and secondary pacakges.
  'test',
]
select = [
  'E',  # pycodestyle
  'F',  # pyflakes
  'UP', # pyupgrade
  # 'D',  # pydocstyle
  "I", # isort
]
ignore = [
  'E501', # line-length. Black handles this.
]
[tool.ruff.per-file-ignores]
'src/glasses/log_parser.py' = ['E999'] # switch statement ignore

[tool.ruff.pydocstyle]
convention = 'google'

[tool.mypy]
disallow_incomplete_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
install_types = true
non_interactive = true
show_column_numbers = true
warn_return_any = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_incomplete_defs = false
disallow_untyped_defs = false


[tool.isort]
profile = "black"
